+++ types/handsontable/index.d.ts
@@ -3,7 +3,7 @@
 // Definitions by: Handsoncode sp. z o.o. <http://handsoncode.net/>, Ryan Riley <https://github.com/panesofglass>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped\
 
-declare namespace ht {
+declare namespace Handsontable {
     interface Options {
         allowEmpty?: boolean;
         allowInsertColumn?: boolean;
@@ -494,19 +494,16 @@ declare namespace ht {
 }
 
 interface Handsontable {
-    new (element: Element, options: ht.Options): ht.Methods;
-    plugins: ht.Plugins;
-    hooks: ht.Hooks;
-    Dom: ht.Dom;
-    dom: ht.Dom;
-    helper: ht.Helper;
-    utils: ht.Utils;
-    renderers: ht.Renderers;
-    editors: ht.Editors;
+    new (element: Element, options: Handsontable.Options): Handsontable.Methods;
+    plugins: Handsontable.Plugins;
+    hooks: Handsontable.Hooks;
+    Dom: Handsontable.Dom;
+    dom: Handsontable.Dom;
+    helper: Handsontable.Helper;
+    utils: Handsontable.Utils;
+    renderers: Handsontable.Renderers;
+    editors: Handsontable.Editors;
 }
 
-declare module "handsontable" {
-    export = Handsontable;
-}
-
-declare var Handsontable: Handsontable;
+export = Handsontable;
+export as namespace Handsontable;
